requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
anyascii==0.3.1
    # via wagtail
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
beautifulsoup4==4.9.3
    # via wagtail
billiard==3.6.4.0
    # via celery
boto3==1.24.82
    # via -r requirements/requirements.in
botocore==1.27.82
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via
    #   -r requirements/requirements.in
    #   django-celery-results
certifi==2022.9.24
    # via
    #   elasticsearch
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==38.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-database-url==1.0.0
    # via -r requirements/requirements.in
django==3.2.15
    # via
    #   -r requirements/requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-comments-xtd
    #   django-contrib-comments
    #   django-extensions
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
    #   wagtail-localize
django-allauth==0.51.0
    # via -r requirements/requirements.in
django-celery-results==2.4.0
    # via -r requirements/requirements.in
django-comments-xtd==2.9.8
    # via -r requirements/requirements.in
django-contrib-comments==2.2.0
    # via django-comments-xtd
django-countries==7.3.2
    # via -r requirements/requirements.in
django-crispy-forms==1.14.0
    # via -r requirements/requirements.in
django-extensions==3.2.1
    # via -r requirements/requirements.in
django-filter==21.1
    # via wagtail
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-storages==1.13.1
    # via -r requirements/requirements.in
django-taggit==2.1.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django-widget-tweaks==1.4.12
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   django-comments-xtd
    #   wagtail
docutils==0.19
    # via django-comments-xtd
draftjs-exporter==2.1.7
    # via wagtail
elasticsearch==7.17.6
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
humanize==4.4.0
    # via -r requirements/requirements.in
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
l18n==2021.3
    # via wagtail
oauthlib==3.2.1
    # via
    #   -r requirements/requirements.in
    #   requests-oauthlib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via redis
pillow==9.1.1
    # via
    #   -r requirements/requirements.in
    #   wagtail
polib==1.1.1
    # via wagtail-localize
prompt-toolkit==3.0.31
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via botocore
python-decouple==3.6
    # via -r requirements/requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2022.2.1
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
redis==4.3.4
    # via -r requirements/requirements.in
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   django-comments-xtd
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
tablib[xls,xlsx]==3.2.1
    # via wagtail
telepath==0.3
    # via wagtail
types-cryptography==3.3.23
    # via pyjwt
typing-extensions==4.3.0
    # via
    #   django-countries
    #   wagtail-localize
urllib3==1.26.12
    # via
    #   botocore
    #   elasticsearch
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==3.0.3
    # via
    #   -r requirements/requirements.in
    #   wagtail-localize
    #   wagtail-seo
wagtail-localize==1.2.1
    # via -r requirements/requirements.in
wagtail-seo==2.2.0
    # via -r requirements/requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
wrapt==1.14.1
    # via deprecated
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via wagtail
xlwt==1.3.0
    # via tablib
